refactor: move set vrange function to operations file
